requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.3
Django==5.0.4
django-celery-beat==2.6.0
django-redis==5.4.0
django-timezone-field==6.1.0
djangorestframework==3.15.1
dnspython==2.6.1
eventlet==0.36.1
greenlet==3.0.3
gunicorn==21.2.0
idna==3.7
inflection==0.5.1
iniconfig==2.0.0
kombu==5.3.6
mirakuru==2.5.2
packaging==24.0
pluggy==1.4.0
port-for==0.7.2
prompt-toolkit==3.0.43
psutil==5.9.8
psycopg2==2.9.9
psycopg2-binary==2.9.9
pytest==8.1.1
pytest-django==4.8.0
pytest-mock==3.14.0
pytest-redis==3.0.2
python-crontab==3.0.0
python-dateutil==2.9.0.post0
PyYAML==6.0.1
redis==5.0.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
